AKSHAT YADAV/CHAT APP/llm-chat-app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
httpx==0.25.2
firebase-admin==6.2.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
streamlit==1.28.2
streamlit-authenticator==0.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
